Edit script.js
Added playRound function. In this function,
the option selected by the player and the computer are compared
with each other, and depending on the selected options,
one of the players is selected as the winner and the other as the loser.
Also, if the player's input is in uppercase or lowercase letters,
or a combination of both, using the "capitalize" helper function,
the first letter is converted to uppercase to match
the option selected by the computer.