Edit script.js
The "game" function was added. The task of this function
is to run the game for 5 rounds and receive input
from the user and the computer and send them to the
"playRound" function to determine the winner in each round.
After each round, each player's score is added to his previous score
and at the end it is determined which player is the winner or loser.