Edit script.js
I realized that the output of the program was not correct, so
I edited the "game" function. Before this, game was played in 5 rounds
and the winner was the player who got the most points.
But this is not correct. Now the game is played in infinite rounds
and whichever of the two players (user or computer) reaches
five points first is the winner. Also, some parts of this function
were changed to make the output look cleaner.